Move OWML back to the general list
